Přesuň přepočítávací funkci do modelu
Místo samostatné stojící funkce přesunuto do modelu DonorsOverview.
